perf(offload): backward prefetch overlap, async grad D2H, bf16 storage default
Three optimizations targeting 1-8B production range where compute/transfer
ratio is 0.5-2.0:

1. Backward weight prefetch: on_backward_layer_end(N) now issues async
   _prefetch_layer(N-1) after eviction, so the H2D transfer overlaps with
   autograd traversal to the next backward call.

2. Async gradient D2H: on_backward_pass_end() now batches all param.grad
   GPU→CPU transfers via submit_d2h() and synchronizes once, instead of
   doing N sequential blocking .cpu() calls.

3. bf16 weight storage default: halves transfer volume with negligible
   precision impact for bf16 training.

All 8 offload modes pass 100-step accuracy validation within 0.025 of
baseline. Throughput improvement is minimal at 130M (compute too fast
for overlap) but scales with model size — the crossover is ~8K
tokens/step on PCIe Gen4.

+++ ironcore/offload/config.py
@@ -50,7 +50,7 @@ class OffloadConfig(BaseConfig):
     # M2: Weight streaming
     weight_offload: bool = False
     weight_prefetch_layers: int = 2
-    weight_storage_precision: str = "fp32"  # precision for streamed weights on host
+    weight_storage_precision: str = "bf16"  # precision for streamed weights on host
     gpu_staging_pool_mb: float = 0.0  # 0 = auto-size based on prefetch_layers * max_layer_bytes
     gpu_staging_chunk_mb: float = 256.0  # chunk size for GPU staging pool
